+++ financial-services-accelerator/components/org.wso2.financial.services.accelerator.common/src/main/java/org/wso2/financial/services/accelerator/common/extension/model/EventRequest.java
@@ -1,0 +1,53 @@
+/**
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ * <p>
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.financial.services.accelerator.common.extension.model;
+
+import org.json.JSONObject;
+
+import java.util.List;
+
+/**
+ * Event Request.
+ */
+public class EventRequest {
+
+    JSONObject payload;
+    List<String> additionalParams;
+
+    public EventRequest(JSONObject payload, List<String> additionalParams) {
+        this.payload = payload;
+        this.additionalParams = additionalParams;
+    }
+
+    public JSONObject getPayload() {
+        return payload;
+    }
+
+    public void setPayload(JSONObject payload) {
+        this.payload = payload;
+    }
+
+    public List<String> getAdditionalParams() {
+        return additionalParams;
+    }
+
+    public void setAdditionalParams(List<String> additionalParams) {
+        this.additionalParams = additionalParams;
+    }
+}
