+++ financial-services-accelerator/components/org.wso2.financial.services.accelerator.common/src/main/java/org/wso2/financial/services/accelerator/common/extension/model/ExternalServiceRequest.java
@@ -1,0 +1,60 @@
+/**
+ * Copyright (c) 2025, WSO2 LLC. (https://www.wso2.com).
+ * <p>
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.financial.services.accelerator.common.extension.model;
+
+/**
+ * Model class to represent the external service request
+ */
+public class ExternalServiceRequest {
+
+    String eventId;
+    Event event;
+    AllowedOperation allowedOperation;
+
+    public ExternalServiceRequest(String eventId, Event event, AllowedOperation allowedOperation) {
+        this.eventId = eventId;
+        this.event = event;
+        this.allowedOperation = allowedOperation;
+    }
+
+    public String getEventId() {
+        return eventId;
+    }
+
+    public void setEventId(String eventId) {
+        this.eventId = eventId;
+    }
+
+    public Event getEvent() {
+        return event;
+    }
+
+    public void setEvent(Event event) {
+        this.event = event;
+    }
+
+    public AllowedOperation getAllowedOperation() {
+        return allowedOperation;
+    }
+
+    public void setAllowedOperation(AllowedOperation allowedOperation) {
+        this.allowedOperation = allowedOperation;
+    }
+
+}
