+++ financial-services-accelerator/components/org.wso2.financial.services.accelerator.identity.extensions/src/main/java/org/wso2/financial/services/accelerator/identity/extensions/auth/extensions/response/handler/impl/FSDefaultResponseTypeHandlerImpl.java
@@ -49,6 +49,10 @@ public class FSDefaultResponseTypeHandlerImpl implements FSResponseTypeHandler {
      */
     public long updateRefreshTokenValidityPeriod(OAuthAuthzReqMessageContext oAuthAuthzReqMessageContext) {
 
+        /**
+         * TODO: call external service and set the refresh token validity period
+         */
+
         return oAuthAuthzReqMessageContext.getRefreshTokenvalidityPeriod();
     }
