+++ financial-services-accelerator/components/org.wso2.financial.services.accelerator.identity.extensions/src/main/java/org/wso2/financial/services/accelerator/identity/extensions/grant/type/handlers/FSPasswordGrantHandler.java
@@ -37,7 +37,7 @@ public class FSPasswordGrantHandler extends PasswordGrantHandler {
         try {
             if (FinancialServicesUtils.isRegulatoryApp(tokReqMsgCtx.getOauth2AccessTokenReqDTO().getClientId())) {
                 OAuth2AccessTokenRespDTO oAuth2AccessTokenRespDTO = super.issue(tokReqMsgCtx);
-                executeInitialStep(oAuth2AccessTokenRespDTO, tokReqMsgCtx);
+                appendParametersToTokenResponse(oAuth2AccessTokenRespDTO, tokReqMsgCtx);
                 tokReqMsgCtx.setScope(IdentityCommonUtils.removeInternalScopes(tokReqMsgCtx.getScope()));
                 return oAuth2AccessTokenRespDTO;
             }
@@ -48,13 +48,17 @@ public class FSPasswordGrantHandler extends PasswordGrantHandler {
     }
 
     /**
-     * Extend this method to perform any actions which requires internal scopes.
+     * Extend this method to perform any actions which requires internal scopes and to append response parameters.
      *
      * @param oAuth2AccessTokenRespDTO
      * @param tokReqMsgCtx
      */
-    public void executeInitialStep(OAuth2AccessTokenRespDTO oAuth2AccessTokenRespDTO,
-                                   OAuthTokenReqMessageContext tokReqMsgCtx) throws IdentityOAuth2Exception {
+    public void appendParametersToTokenResponse(OAuth2AccessTokenRespDTO oAuth2AccessTokenRespDTO,
+                                                OAuthTokenReqMessageContext tokReqMsgCtx)
+            throws IdentityOAuth2Exception {
 
+        /**
+         * TODO: call external service and set the required response parameters
+         */
     }
 }
